refactor(kernel): organize journal support modules
