+++ packages/kernel/src/store/daemon-runtime-queue.ts
@@ -5,8 +5,8 @@ import type { WriteError } from "../domain/index.ts";
 import type { FlushReport, WriteOp } from "../ports/write-coordinator.ts";
 import type { VcsCommitAuthor } from "../ports/version-control-system.ts";
 import type { WriteAttribution } from "../schemas/actor-attribution.ts";
-import type { OperationalActor } from "./write-journal-types.ts";
-import { singleWriteIntegrityDomain, type WriteIntegrityDomain } from "./write-integrity-domain.ts";
+import type { OperationalActor } from "../write-coordination/journal/types.ts";
+import { singleWriteIntegrityDomain, type WriteIntegrityDomain } from "../write-coordination/journal/integrity-domain.ts";
 import type { makeJournaledWriteCoordinator } from "./write-journal-coordinator.ts";
 
 export type DaemonWritePriority = "interactive" | "normal" | "background" | "maintenance";
