+++ packages/kernel/src/store/daemon-runtime.ts
@@ -30,9 +30,9 @@ import {
 } from "./daemon-runtime-authority-publication.ts";
 import type { DaemonDrainOptions, DaemonMaterializerBatchOptions } from "./daemon-runtime-options.ts";
 export type { DaemonDrainOptions, DaemonMaterializerBatchOptions } from "./daemon-runtime-options.ts";
-import { acquireDaemonGlobalLock, assertDaemonGlobalLockHeld, type DaemonGlobalLock } from "./write-journal-locks.ts";
+import { acquireDaemonGlobalLock, assertDaemonGlobalLockHeld, type DaemonGlobalLock } from "../write-coordination/journal/locks.ts";
 import { makeJournaledWriteCoordinator, makeOperationalJournaledWriteCoordinator, recoverJournaledWrites } from "./write-journal-coordinator.ts";
-import type { OperationalActor } from "./write-journal-types.ts";
+import type { OperationalActor } from "../write-coordination/journal/types.ts";
 import { writeOpTouchedPaths } from "./write-journal-operations.ts";
 import {
   createDaemonProjectionGenerationManager,
