+++ packages/kernel/src/store/declared-entity-preconditions.ts
@@ -5,7 +5,7 @@ import type {
 import type { HarnessLayoutInput } from "../layout/index.ts";
 import type { WriteOp } from "../ports/write-coordinator.ts";
 import { documentTargetPath } from "./write-journal-operations-internal.ts";
-import { rejectWrite } from "./write-journal-rejection.ts";
+import { rejectWrite } from "../write-coordination/journal/rejection.ts";
 
 export function declaredEntityPreconditions(
   payload: DeclaredEntityDocumentWritePayload
