+++ packages/kernel/src/store/write-journal-coordinator.ts
@@ -21,16 +21,16 @@ import {
 } from "../layout/index.ts";
 import type { ProjectionChangeEvent } from "../projection/projection-change-event.ts";
 import { captureAuthoredProjectionFingerprint } from "../projection/projection-source-baseline.ts";
-import { appendJsonLineDurably, readDurableState, readPayloadRef, writeWatermarkDurably, writeFileDurably } from "./write-journal-durable.ts";
-import { assertCommitPlanAddable, commitTouchedPaths } from "./write-journal-git.ts";
+import { appendJsonLineDurably, readDurableState, readPayloadRef, writeWatermarkDurably, writeFileDurably } from "../write-coordination/journal/durable.ts";
+import { assertCommitPlanAddable, commitTouchedPaths } from "../write-coordination/journal/publication/git.ts";
 import { makeLocalVersionControlSystem } from "../persistence/git/local-version-control-system.ts";
-import { assertCodeDocGitEvidence, assertNoUncoordinatedCodeDocChange } from "./write-journal-code-doc-policy.ts";
-import { writeJournalRecordCommitSummary } from "./write-journal-commit-summary.ts";
+import { assertCodeDocGitEvidence, assertNoUncoordinatedCodeDocChange } from "../write-coordination/journal/operations/code-doc-policy.ts";
+import { writeJournalRecordCommitSummary } from "../write-coordination/journal/publication/commit-summary.ts";
 import { runLedgerMaterializer } from "./ledger-materializer.ts";
 import { createAttributionEvent, makeLocalGitAttributionEventStore, planAttributionEventCommit, type AttributionEventStore } from "../write-coordination/attribution/legacy-attribution-event-store.ts";
-import { assertDirectWriteAllowed, withRepoLocks, WriteLockHeldError } from "./write-journal-locks.ts";
-import { NonTaskWriteEntityError, taskIdForJournalRecord } from "./write-journal-entity.ts";
-import { rejectWrite, WriteRejectedError } from "./write-journal-rejection.ts";
+import { assertDirectWriteAllowed, withRepoLocks, WriteLockHeldError } from "../write-coordination/journal/locks.ts";
+import { NonTaskWriteEntityError, taskIdForJournalRecord } from "../write-coordination/journal/operations/entity.ts";
+import { rejectWrite, WriteRejectedError } from "../write-coordination/journal/rejection.ts";
 import {
   assertRecordMatchesAttributedOp,
   assertRecordMatchesOperationalOp,
@@ -38,7 +38,7 @@ import {
   createOperationalJournalRecord,
   decodeWriteAttribution,
   uniquePendingRecords
-} from "./write-journal-records.ts";
+} from "../write-coordination/journal/records.ts";
 import {
   applyWriteOp,
   documentWritesForWriteOp,
@@ -46,13 +46,13 @@ import {
   validateWriteTransaction,
   writeOpTouchedPaths
 } from "./write-journal-operations.ts";
-import { reconcileDurableFlush, shouldWaitForForeignCommitter } from "./write-journal-receipt.ts";
-import { semanticCommitMessage } from "./write-journal-authority-trailer.ts";
-import { recoverJournalIntegrityDomains } from "./write-journal-domain-recovery.ts";
-import { recordsForWriteIntegrityDomain, singleWriteIntegrityDomain } from "./write-integrity-domain.ts";
-import { memoizePublicationVcs } from "./write-journal-publication-vcs.ts";
-import { rebuildProjectionHash } from "./write-journal-projection-publication.ts";
-import type { ApplyMarkerRecord, DeleteAuditRecord, JournaledWriteCoordinatorOptions, JournalRecoveryOptions, LockConflictRetryOptions, LockTakeoverRecord, OperationalActor, OperationalJournaledWriteCoordinatorOptions, ReadableJournalRecord, WriteWatermark } from "./write-journal-types.ts";
+import { reconcileDurableFlush, shouldWaitForForeignCommitter } from "../write-coordination/journal/receipt.ts";
+import { semanticCommitMessage } from "../write-coordination/journal/publication/authority-trailer.ts";
+import { recoverJournalIntegrityDomains } from "../write-coordination/journal/recovery/integrity-domains.ts";
+import { recordsForWriteIntegrityDomain, singleWriteIntegrityDomain } from "../write-coordination/journal/integrity-domain.ts";
+import { memoizePublicationVcs } from "../write-coordination/journal/publication/memoized-vcs.ts";
+import { rebuildProjectionHash } from "../write-coordination/journal/publication/projection.ts";
+import type { ApplyMarkerRecord, DeleteAuditRecord, JournaledWriteCoordinatorOptions, JournalRecoveryOptions, LockConflictRetryOptions, LockTakeoverRecord, OperationalActor, OperationalJournaledWriteCoordinatorOptions, ReadableJournalRecord, WriteWatermark } from "../write-coordination/journal/types.ts";
 export type {
   JournalActor,
   JournalRecordV1,
@@ -62,7 +62,7 @@ export type {
   LockConflictRetryOptions,
   OperationalActor,
   ReadableJournalRecord
-} from "./write-journal-types.ts";
+} from "../write-coordination/journal/types.ts";
 
 const defaultOperationalActor: OperationalActor = { scope: "operational", kind: "agent", id: "write-coordinator" };
 // Flush writes the full op-id set before compaction for recovery safety, then
