+++ packages/kernel/src/write-coordination/journal/integrity-domain.ts
