+++ packages/kernel/test/store/authority-attribution-event-v2-log.test.ts
@@ -76,7 +76,7 @@ test("immutable V2 shard converges after SIGKILL at every publication boundary",
       const bytes = encodeAuthorityAttributionEventV2Bytes(event);
       const child = spawnSync(process.execPath, [
         "--experimental-strip-types", "--eval",
-        "import { appendImmutableBytesDurably } from './packages/kernel/src/store/write-journal-durable.ts'; appendImmutableBytesDurably(process.env.EVENT_PATH, Buffer.from(process.env.EVENT_BYTES, 'base64'));"
+        "import { appendImmutableBytesDurably } from './packages/kernel/src/write-coordination/journal/durable.ts'; appendImmutableBytesDurably(process.env.EVENT_PATH, Buffer.from(process.env.EVENT_BYTES, 'base64'));"
       ], {
         cwd: process.cwd(),
         env: {
