+++ packages/kernel/test/store/same-task-fifo.test.ts
@@ -8,7 +8,7 @@ import path from "node:path";
 import { Effect } from "effect";
 import { hashTaskProjectionRows, rebuildTaskProjection, sha256Text } from "../../src/index.ts";
 import { makeJournaledWriteCoordinator } from "../../src/store/index.ts";
-import { resolveCommitPlan } from "../../src/store/write-journal-git.ts";
+import { resolveCommitPlan } from "../../src/write-coordination/journal/publication/git.ts";
 import { moduleEntityId, taskEntityId } from "../../src/domain/index.ts";
 import { docWrite, withTempStore } from "./helpers.ts";
