+++ tools/check-bypass-write-boundary.mjs
@@ -9,6 +9,7 @@ import { entryValues, loadGateAllowlist } from "./gate-allowlists/load-gate-allo
 const targetRoots = [
   "packages/kernel/src/store",
   "packages/kernel/src/persistence",
+  "packages/kernel/src/write-coordination",
   "packages/adapters/local/src",
   "packages/daemon/src",
   "packages/cli/src/commands"
